perf(reader): improve aperture edge check routine
This commit improves the routine for pulling the apertures and doors inside if they are touching the edge of the parent face.

The new approach pulls the vertices inside perpendicular to the edge if they are touching the edge and it offsets them inwards if they are touching a corner.